requirements.txt
# CSC110 Fall 2020: Python libraries in Final Project.

# Testing and code checking
hypothesis
pytest
python-ta

# Data Reapading
os
csv
json
ast

# Graphics and data visualization
plotly.express
pygame==2.0.0.dev10

# User interface
tkinter
PIL

psutilPIL
